more conversion of xrd to main lib
